+++ common/gradle.properties
@@ -8,6 +8,9 @@ cloth_config_version = 11.1.106
 # https://modrinth.com/mod/sodium/versions
 sodium_version = mc1.20.1-0.5.13-fabric
 
+# https://modrinth.com/mod/alexs-caves/versions
+alexscaves_version = 2.0.2
+
 # https://modrinth.com/mod/tis3d/version/MC1.20.1-fabric-1.7.7
 tis3d_version = MC1.20.1-fabric-1.7.7
 immptl_version=v3.3.9-mc1.20.1
