+++ common/src/main/resources/valkyrienskies-common.mixins.json
@@ -55,7 +55,6 @@
     "feature.entity_fall_particles.MixinLivingEntity",
     "feature.entity_movement_packets.MixinServerEntity",
     "feature.explosions.MixinExplosion",
-    "feature.explosions.MixinNuclearExplosion",
     "feature.fire_between_ship_and_world.LavaFluidMixin",
     "feature.fluid_escaping_ship_config.MixinFlowingFluid",
     "feature.fluid_escaping_ship_config.MixinVineBlock",
@@ -101,6 +100,7 @@
     "feature.world_weather.MixinBiome",
     "feature.world_weather.MixinMeltingBlocks",
     "feature.world_weather.MixinServerLevel",
+    "mod_compat.alex_caves.MixinNuclearExplosion",
     "mod_compat.bluemap.HiresTileModelAccessor",
     "mod_compat.bluemap.MixinBmMap",
     "mod_compat.bluemap.MixinHiresModelManager",
